Refactor: extract more types from Core.Shake
